+++ prov/terraform/openstack/osinstance.go
@@ -394,13 +394,13 @@ func addResources(ctx context.Context, opts osInstanceOptions, fipAssociateName,
 		"/attributes/private_address")] = privateIPKey
 
 	// Get connection info (user, private key)
-	user, privateKey, err := commons.GetConnInfoFromEndpointCredentials(ctx, opts.deploymentID, opts.nodeName)
+	user, pk, err := commons.GetConnInfoFromEndpointCredentials(ctx, opts.deploymentID, opts.nodeName)
 	if err != nil {
 		return err
 	}
 
-	return commons.AddConnectionCheckResource(opts.infrastructure, user,
-		privateKey, accessIP, instance.Name, env)
+	return commons.AddConnectionCheckResource(ctx, opts.deploymentID, opts.nodeName, opts.infrastructure, user,
+		pk, accessIP, instance.Name, env)
 
 }
 
@@ -418,7 +418,7 @@ func computeFloatingIPAddress(ctx context.Context, opts osInstanceOptions,
 	resultChan := make(chan string, 1)
 	go func() {
 		for {
-			if fip, _ := deployments.GetInstanceCapabilityAttributeValue(ctx, deploymentID, networkNodeName, instanceName, "endpoint", "floating_ip_address"); fip != nil && fip.RawString() != "" {
+			if fip, _ := deployments.GetInstanceCapabilityAttributeValue(ctx, deploymentID, networkNodeName, "0", "endpoint", "floating_ip_address"); fip != nil && fip.RawString() != "" {
 				resultChan <- fip.RawString()
 				return
 			}